Fix Project Config
