+++ client/scripts/init.sh
@@ -27,7 +27,3 @@ cat $CONFIG_FILE
 # Load Rundeck Information (Projects, Users, etc.)
 ./bin/cli load --rundeck_url $RUNDECK_URL --config_file "$CONFIG_FILE" --path /rundeck-cli
 ./bin/cli addUsers --rundeck_url $RUNDECK_URL --config_file "$CONFIG_FILE" --path /rundeck-cli
-
-sleep 10
-
-./bin/cli updateProject --rundeck_url $RUNDECK_URL --config_file "$CONFIG_FILE" --path /rundeck-cli
